refactor: modularize responsibilities
